Added more tests and caught a really bad bug in the hasher.
Some details about the bug. Which was discovered via testing. And it was really bad. So this is money well spent.

We used the original image size % 5 as a hash, rather than the resized image. So basically, detection across different sizes was broken. This might explain many of the duplicate images from 9gag, because I think 9 gag does proper image processing and standardizes their images. Testing well done.